🐞 Fix problem with card links on initial load
Some cards are supposed to change links based on which
language is selected. However, this didn't work properly.
Turns out the problem was that the choice of language
wasn't available client-side.

This commits fixes this problem by checking the language
cookie server-side as well.

The implementation is, unfortunately, not perfect. The
first time the user visits the site, the default language
is used. On subsequent refreshes, the correct language
is used.